Delete introduction
For this project, i create a "word cloud" from a text by writing a script. This script process the text, remove punctuation, ignore case and words that do not contain all alphabets, count the frequencies, and ignore uninteresting or irrelevant words. A dictionary is the output of the calculate_frequencies function. The wordcloud module will then generate the image from my dictionary.